// src/common/enums/daily-expense.enum.ts
export enum DailyExpenseType {
  ROOM = 'room',
  OWN = 'own',
}

export enum DailyExpenseStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
}

export enum DailyExpenseCategory {
  GROCERY = 'GROCERY',
  VEGETABLES = 'VEGETABLES',
  DAIRY = 'DAIRY',
  GAS_CYLINDER = 'GAS_CYLINDER',
  CLEANING = 'CLEANING',
  WATER = 'WATER',
  VAKIL_MASI = 'Vakil/Masi',
  UTILITIES = 'UTILITIES',
  PERSONAL = 'PERSONAL',
  MEDICINE = 'MEDICINE',
  TRAVEL = 'TRAVEL',
  OTHER = 'OTHER',
}